move api interactions (get, save) to new uploaders.py, rename uploaders to transformers
